refactor(sys): modify stub
